_old/requirements-enterprise.txt
# requirements-enterprise.txt - Dependencias para sistema enterprise
# Ubicación: C:\TradingBot_v10\requirements-enterprise.txt

# Dependencias base del trading bot
-r requirements.txt

# Dependencias para sistema enterprise
# =====================================

# Configuración y utilidades
PyYAML>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
dataclasses-json>=0.6.0

# Base de datos y almacenamiento
asyncpg>=0.28.0
redis>=4.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Streaming y mensajería
kafka-python>=2.0.2
confluent-kafka>=2.2.0
websockets>=11.0.0
aiohttp>=3.8.0

# Procesamiento de datos
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
ta-lib>=0.4.0

# Monitoreo y métricas
prometheus-client>=0.17.0
grafana-api>=1.0.3
psutil>=5.9.0

# Seguridad y encriptación
cryptography>=41.0.0
pycryptodome>=3.18.0
hashicorp-vault>=0.1.0
requests>=2.31.0

# Logging y auditoría
structlog>=23.1.0
python-json-logger>=2.0.0

# Testing y desarrollo
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentación
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Utilidades del sistema
click>=8.1.0
rich>=13.5.0
tqdm>=4.65.0
schedule>=1.2.0

# Análisis de datos
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning avanzado
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
optuna>=3.3.0

# APIs y servicios web
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.2.0

# Manejo de archivos y compresión
pyarrow>=12.0.0
parquet>=1.3.1
lz4>=4.3.0

# Networking y comunicación
httpx>=0.24.0
websocket-client>=1.6.0
aiofiles>=23.2.0

# Configuración de desarrollo
pre-commit>=3.3.0
isort>=5.12.0
bandit>=1.7.0
safety>=2.3.0

# Dependencias específicas de Windows
pywin32>=306; sys_platform == "win32"
wmi>=1.5.1; sys_platform == "win32"

# Dependencias opcionales para producción
gunicorn>=21.2.0
supervisor>=4.2.0